% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.TARMA.R
\name{print.TARMA}
\alias{print.TARMA}
\alias{coef.TARMA}
\alias{vcov.TARMA}
\alias{residuals.TARMA}
\title{Methods for TARMA fits}
\usage{
\method{print}{TARMA}(x, digits = max(3L, getOption("digits") - 3L), se = TRUE, ...)

\method{coef}{TARMA}(object, ...)

\method{vcov}{TARMA}(object, ...)

\method{residuals}{TARMA}(object, ...)
}
\arguments{
\item{x}{A \code{TARMA} fit.}

\item{digits}{Number of decimal digits for the output.}

\item{se}{Logical. if \code{TRUE} (the default) prints also the standard errors.}

\item{\dots}{Further parameters.}

\item{object}{A \code{TARMA} fit.}
}
\value{
No return value, called for side effects
}
\description{
Methods for TARMA fits
}
\seealso{
\code{\link{TARMA.fit}} and \code{\link{TARMA.fit2}} for TARMA modelling, \code{\link{plot.tsfit}} for plotting TARMA fits and forecasts.
}
