% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.TARMAtest.R
\name{print.TARMAtest}
\alias{print.TARMAtest}
\title{Methods for TARMA tests}
\usage{
\method{print}{TARMAtest}(x, ...)
}
\arguments{
\item{x}{A \code{TARMAtest} object.}

\item{\dots}{Further parameters passed to \code{\link{print.htest}}.}
}
\value{
No return value, called for side effects
}
\description{
Methods for TARMA tests
}
\details{
Print method for \code{TARMAtest} objects. Prints the results using the method for class \code{htest}
and adds critical values extracted from \code{\link{ACValues}} for the test for threshold nonlinearity
and \code{\link{supLMQur}} for the unit root test against the TARMA alternative.
Note that the bootstrap version of the tests also print the bootstrap p-value.
}
\seealso{
\code{\link{print.htest}}
}
\author{
Simone Giannerini, \email{simone.giannerini@unibo.it}

Greta Goracci, \email{greta.goracci@unibz.it}
}
