% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supLMQur.R
\docType{data}
\name{supLMQur}
\alias{supLMQur}
\title{Tabulated Critical Values for the Unit Root IMA vs TARMA test}
\format{
\subsection{\code{supLMQur}}{

A 4-dimensional array that contains 4 critical values (at level 10\%, 5\%, 1\%, 0.1\%)
for each combination of
\describe{
\item{\code{pa}}{Lower bound for the threshold range. From 0.01 to 0.4}
\item{\code{th}}{MA(1) parameter.}
\item{\code{n}}{Sample size.}
}
}
}
\source{
\insertRef{Cha24}{tseriesTARMA}
}
\usage{
supLMQur
}
\description{
\loadmathjax
The data provides asymptotic null critical
values for the unit root supLM test described in \insertCite{Cha24}{tseriesTARMA}.
They are used with the following tests:
\itemize{
\item \code{\link{TARMAur.test}}
\item \code{\link{TARMAur.test.B}}
}

Provided \code{pb = 1- pa}.
}
\keyword{datasets}
