% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tseriesTARMA.R
\docType{package}
\name{tseriesTARMA}
\alias{tseriesTARMA-package}
\alias{tseriesTARMA}
\title{tseriesTARMA: Analysis of Nonlinear Time Series through Threshold Autoregressive Moving Average Models (TARMA) models}
\description{
It provides advanced functions for:
\itemize{
\item TARMA model fitting and forecasting:
\itemize{
\item Least Squares fitting of a full subset TARMA model, including robust LS fitting.
\item Maximum Likelihood fitting of a subset TARMA model with common MA parts and possible covariates.
}
\item TARMA testing for threshold type nonlinearity:
\itemize{
\item Tests for AR vs TAR (asymptotic, bootstrap, wild bootstrap)
\item Tests for ARMA vs TARMA with both i.i.d. errors and GARCH errors.
}
\item Unit-root testing against a stationary TARMA model
}
}
\references{
\itemize{
\item \insertRef{Gia23}{tseriesTARMA}
\item \insertRef{Ang23}{tseriesTARMA}
\item \insertRef{Gia22}{tseriesTARMA}
\item \insertRef{Gia21}{tseriesTARMA}
\item \insertRef{Gor21}{tseriesTARMA}
\item \insertRef{Gor23}{tseriesTARMA}
\item \insertRef{Cha19}{tseriesTARMA}
\item \insertRef{Cha24}{tseriesTARMA}
}
}
\author{
Simone Giannerini, \email{simone.giannerini@uniud.it}

Greta Goracci, \email{greta.goracci@unibz.it}
}
\keyword{internal}
